Refactor `loader.py` to `/loaders`
