+++ src/yads/__init__.py
@@ -1,4 +1,4 @@
-from .loader import from_yaml, from_string, from_dict
+from .loaders import from_yaml, from_string, from_dict
 
 __all__ = [
     "from_yaml",
