+++ src/yads/loaders/common/__init__.py
@@ -1,0 +1,5 @@
+"""Common building utilities for yads loaders."""
+
+from .spec_builder import SpecBuilder
+
+__all__ = ["SpecBuilder"]
